requirements.txt
Flask==2.2.2
flasgger==0.9.5
numpy==1.23.4
pandas==1.5.1
Flask-SQLAlchemy==3.0.2
nltk==3.8.1


